Move local state setup function
