+++ tests/test_has_code.py
@@ -1,6 +1,6 @@
 import pytest
 import tests.helper as helper
-from pythonwhat.local import setup_state
+from pythonwhat.test_exercise import setup_state
 
 
 @pytest.mark.parametrize("stu, passes", [("", False), ("c", False), ("a == c", True)])
